+++ test/legion/recovery_test.exs
@@ -118,7 +118,7 @@ defmodule Legion.RecoveryTest do
 
     test_pid = self()
 
-    stub(ReqLLM, :generate_object, fn _model, _messages, _schema ->
+    stub(ReqLLM, :generate_object, fn _model, _messages, _schema, _opts ->
       send(test_pid, {:recovering, self()})
 
       receive do
@@ -174,7 +174,7 @@ defmodule Legion.RecoveryTest do
 
     test_pid = self()
 
-    stub(ReqLLM, :generate_object, fn _model, _messages, _schema ->
+    stub(ReqLLM, :generate_object, fn _model, _messages, _schema, _opts ->
       send(test_pid, {:recovering, self()})
 
       receive do
@@ -212,7 +212,7 @@ defmodule Legion.RecoveryTest do
 
     test_pid = self()
 
-    stub(ReqLLM, :generate_object, fn _model, _messages, _schema ->
+    stub(ReqLLM, :generate_object, fn _model, _messages, _schema, _opts ->
       send(test_pid, {:recovering, self()})
 
       receive do